Remove redundant JettyRPCFramework abstraction
